requirements.txt
aws-cdk-lib==2.34.2
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-alpha==2.34.2a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.34.2a0
aws-cdk.aws_apigatewayv2_authorizers_alpha==2.34.2a0